dev/api/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.1
flatbuffers==24.3.25
gast==0.4.0
git-filter-repo==2.45.0
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
idna==3.7
ipykernel==6.23.1
# ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
